move RenderCheckContainer
