+++ src/dashboard/DashHome.js
@@ -4,11 +4,9 @@ import {
   GANotActiveLogin,
   EngagedUser
 } from "./GoogleAnalytics/index";
-import GraphContainer from "../GraphContainer";
 import { getToken, decodeToken } from "./auth/Auth";
 import CreateAccount from "./CreateAccount";
 import AccountHandler from "../dashboard/DashboardAccount/AccountHandler";
-import LandingPage from "./LandingPage";
 
 function DashHome() {
   const signedIn = getToken();
